pyproject.toml
[project]
name = "tmux-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.64.0",
    "google-generativeai>=0.8.5",
    "litellm>=1.75.8",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]
